\name{agreement-package}
\alias{agreement-package}
\docType{package}
\title{
agreement: Analyse the agreement between two measurement methods
}
\description{
The agreement is a package to investigate agreement
between two measurement methods using a simulation
approach. This approach is the same as that presented
in literature.
}
\details{
\tabular{ll}{
Package: \tab agreement\cr
Type: \tab Package\cr
LazyData: \tab yes\cr
LazyLoad: \tab yes\cr
Encoding: \tab latin1\cr
Depends: \tab R (>= 2.7.0), stats, utils, tools\cr
Version: \tab 1.0-1\cr
Date: \tab 2008-09-18\cr
License: \tab GPL-2\cr
URL: \tab http://www.r-project.org/, http://cran.r-project.org/\cr
}

Index:
\preformatted{
lin.simulation          Performs a summary of the simulation
}

Further information is available in the following vignettes:
\tabular{ll}{
\code{agreement} \tab agreement: Analyse the agreement between two measurement methods (source, pdf)\cr
}
}
\author{
Fabio Frascati, Bruno Mario Cesana

Maintainer: Fabio Frascati <fabio.frascati@studenti.unimi.it>
}
\keyword{ package }
