% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefficients.R
\name{calc_all}
\alias{calc_all}
\title{Calculate all chance-adjusted indexes of categorical agreement}
\usage{
calc_all(.data, categories = NULL, weighting = c("identity", "linear",
  "quadratic"), format = c("long", "wide"))
}
\arguments{
\item{.data}{Required.}

\item{categories}{Optional.}

\item{weighting}{Optional.}

\item{format}{Optional.}
}
\value{
A tibble containing the estimates of all chance-adjusted indexes of
categorical agreement.
}
\description{
This function calculates all chance-adjusted indexes of categorical agreement
(i.e., S, gamma, kappa, pi, and alpha) and then binds their results together
into a single tibble. All indexes are calculated using the same data,
categories and weighting scheme, so they are directly comparable.
}
\examples{
fish <- data("fish")
calc_all(fish, categories = c(1, 2, 3), weighting = "identity")
}
