% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefficients.R
\name{calc_alpha}
\alias{calc_alpha}
\title{Calculate the alpha coefficient using the generalized formula}
\usage{
calc_alpha(.data, categories = NULL, weighting = c("identity",
  "linear", "quadratic"))
}
\arguments{
\item{.data}{Required.}

\item{categories}{Optional.}

\item{weighting}{Optional.}
}
\value{
A 1x4 tibble with the following columns: \item{Approach}{The name of
the coefficient} \item{Observed}{Percent of agreement observed in the data}
\item{Expected}{Percent of agreement expected by chance (given the
coefficient's assumptions)} \item{Adjusted}{The estimated chance-adjusted
agreement (i.e., the alpha coefficient itself)}
}
\description{
Krippendorff's alpha coefficient is a chance-adjusted index of categorical
agreement. It estimates chance agreement using distribution-based
assumptions. This function calculates this coefficient using a generalized
formula that can accommodate multiple raters, multiple categories, and
ordered or unordered categories (through a weighting scheme).
}
\references{
\url{https://doi.org/10/b6v5kt}
}
