% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlations.R
\name{calc_icc}
\alias{calc_icc}
\title{Calculate various intraclass correlation coefficients}
\usage{
calc_icc(.data, format = c("long", "wide"), digits = 3)
}
\arguments{
\item{.data}{Required.}

\item{format}{Optional.}

\item{digits}{Optional.}
}
\value{
A tibble
}
\description{
Intraclass correlation coefficients...
}
\references{
\url{https://doi.org/10/br5ffs}
}
