% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefficients.R
\name{calc_s}
\alias{calc_s}
\title{Calculate the S score using the generalized formula}
\usage{
calc_s(.data, categories = NULL, weighting = c("identity", "linear",
  "quadratic"), format = c("long", "wide"))
}
\arguments{
\item{.data}{Required.}

\item{categories}{Optional.}

\item{weighting}{Optional.}

\item{format}{Optional.}
}
\value{
A tibble containing the percent of agreement observed in the data,
the percent of agreement expected by chance, and the chance-adjusted index
of reliability (i.e., the coefficient itself). These numbers are all
calculated under the assumptions of the coefficient.
}
\description{
Bennett et al.'s S score is a chance-adjusted index of categorical agreement.
It estimates chance agreement using category-based assumptions. This function
calculates this coefficient using a generalized formula that can accommodate
multiple raters, multiple categories, and ordered or unordered categories
(through a weighting scheme).
}
\references{
\url{https://doi.org/10/brfm77}
}
