% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bennett_s.R
\name{cat_s}
\alias{cat_s}
\title{Calculate the S Score}
\usage{
cat_s(.data, categories = NULL, weighting = c("identity", "linear",
  "quadratic"), bootstrap = 2000, interval = 0.95, digits = 3,
  details = FALSE, warnings = TRUE)
}
\arguments{
\item{.data}{\emph{Required.} An matrix or data frame where rows correspond to
objects of measurement, columns correspond to sources of measurement (e.g.,
raters), and values correspond to category assignments. Note that \code{NA}
will be treated as missing values, so if \code{NA} is a meaningful code in
your data, then recode them to another value and reserve \code{NA} for
missing values.}

\item{categories}{\emph{Optional.} A numeric vector of possible categories. Useful
when \code{.data} may not contain all possible categories.}

\item{weighting}{\emph{Optional.} A string indicating whether to use "identity"
weights for unordered categories or "linear" or "quadratic" weights for
ordered categories (default = "identity").}

\item{bootstrap}{\emph{Optional.} A positive integer specifying how many bootstrap
resamples to use in calculating the confidence intervals (default = 2000).
Or, to prevent bootstrapping, set to \code{NULL}.}

\item{interval}{\emph{Optional.} A real number greater than 0 and less than 1
specifying the width of the confidence interval to calculate such as 0.95
for a 95\% confidence interval (default = 0.95).}

\item{digits}{\emph{Optional.} A positive integer specifying how many digits to
round the specific agreement estimates to (default = 3). Or, to prevent
rounding, set to \code{NULL}.}

\item{details}{\emph{Optional.} A logical specifying whether to include additional
columns in the output detailing the number of objects, raters, categories,
valid bootstrap resamples, and interval width (default = \code{FALSE}).}

\item{warnings}{\emph{Optional.} A logical specifying whether to include warnings
(default = \code{TRUE}).}
}
\value{
A tibble containing the Approach (i.e., S), the weighting scheme
used, the percent agreement observed in the data (unadjusted), the percent
agreement expected based on chance, and the chance-adjusted index (i.e., S
score). If \code{bootstrap} is used, the lower and upper bounds of a
confidence interval for the chance-adjusted index are also provided. If
\code{details} is \code{TRUE}, additional detail columns are provided.
}
\description{
Bennett et al.'s S score is a chance-adjusted index of categorical agreement.
It estimates chance agreement using category-based assumptions. This function
calculates this coefficient using a generalized formula that can accommodate
multiple raters, multiple categories, and ordered or unordered categories
(through a weighting scheme).
}
\examples{
data(unordered)
cat_s(unordered)
data(ordered)
cat_s(ordered, weighting = "linear")
}
\references{
Bennett, E. M., Alpert, R., & Goldstein, A. C. (1954).
Communication through limited response questioning. \emph{The Public Opinion
Quarterly, 18}(3), 303-308. \url{https://doi.org/10/brfm77}

Gwet, K. L. (2014). \emph{Handbook of inter-rater reliability: The
definitive guide to measuring the extent of agreement among raters} (4th
ed.). Gaithersburg, MD: Advanced Analytics.
}
\seealso{
Other agreement functions for categorical data: \code{\link{cat_specific}}
}
\concept{agreement functions for categorical data}
