% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorical_specific_agreement.R
\name{cat_specific}
\alias{cat_specific}
\title{Calculate Category-Specific Agreement}
\usage{
cat_specific(.data, categories = NULL, bootstrap = 2000,
  warnings = TRUE)
}
\arguments{
\item{.data}{\emph{Required.} A matrix or data frame containing categorical data
where each row corresponds to a single object of measurement (e.g.,
subject) and each column corresponds to a single source of measurement
(e.g., rater). Cells should contain numbers or characters indicating the
discrete category that the corresponding rater assigned the corresponding
object to. Cells should contain \code{NA} if a particular assignment is
missing (e.g., that object was not assigned to a category by that rater).}

\item{categories}{\emph{Optional.} A vector (numeric, character, or factor)
containing all possible categories that objects could have been assigned
to. When this argument is omitted or set to \code{NULL}, the possible
categories are assumed to be those observed in \code{.data}. However, in
the event that not all possible categories are observed in \code{.data},
this assumption may be misleading and so the possible categories, and their
ordering, can be explicitly specified. (default = NULL)}

\item{bootstrap}{\emph{Optional.} A single non-negative integer that specifies how
many bootstrap resamplings should be computed (used primarily for
estimating confidence intervals and visualizing uncertainty). To skip
bootstrapping, set this argument to 0. (default = 2000)}

\item{warnings}{\emph{Optional.} A single logical value that specifies whether
warnings should be displayed. (default = TRUE).}
}
\value{
An object of type 'spa' containing the results and details.
\describe{\item{observed}{A named numeric vector containing the observed
specific agreement for each category} \item{boot_results}{A list containing
the results of the bootstrapping procedure} \item{details}{A list
containing the details of the analysis, such as the formatted \code{codes},
relevant counts, weighting scheme and weight matrix.} \item{call}{The
function call that created these results.}}
}
\description{
Specific agreement is an index of the reliability of categorical
measurements. It describes the amount of agreement observed with regard to
each possible category. With two raters, the interpretation of specific
agreement for any category is the probability of one rater assigning an item
to that category given that the other rater has also assigned that item to
that category. With more than two raters, the interpretation becomes the
probability of a randomly chosen rater assigning an item to that category
given that another randomly chosen rater has also assigned that item to that
category. When applied to binary (i.e., dichotomous) data, specific agreement
on the positive category is often referred to as positive agreement (PA) and
specific agreement on the negative category is often referred to as negative
agreement (NA). In this case, PA is equal to the F1 score frequently used in
computer science.
}
\references{
Dice, L. R. (1945). Measures of the amount of ecologic
association between species. \emph{Ecology, 26}(3), 297-302.
\url{https://doi.org/10/dsb8pd}

Fleiss, J. L. (1975). Measuring agreement between two judges on the presence or absence of a trait. \emph{Biometrics, 31}(3), 651-659. \url{https://doi.org/10/fxdb8x}

Uebersax, J. S. (1982). A design-independent method for measuring
the reliability of psychiatric diagnosis. \emph{Journal of Psychiatric Research,
17}(4), 335-342. \url{https://doi.org/10/fbbdfn}

Cicchetti, D. V., & Feinstein, A. R. (1990). High agreement but
low kappa: II. Resolving the paradoxes. \emph{Journal of Clinical Epidemiology,
43}(6), 551-558. \url{https://doi.org/10/czkxkb}
}
\seealso{
Other functions for categorical data: \code{\link{cat_adjusted}}
}
\concept{functions for categorical data}
\concept{functions for specific agreement}
