% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorical_oop.R
\name{confint.agreement_cai}
\alias{confint.agreement_cai}
\title{Calculate confidence intervals for chance-adjusted agreement indexes}
\usage{
\method{confint}{agreement_cai}(object, level = 0.95, type = c("bca", "perc", "basic", "norm"), ...)
}
\arguments{
\item{object}{A chance-adjusted agreement or "cai" object.}

\item{level}{A single real number between 0 and 1 that indicates the
confidence level or width of the confidence intervals.}

\item{type}{A single string indicating which type of bootstrap confidence
interval to calculate. Options currently include "bca" for bias-corrected
and accelerated bootstrap CIs, "perc" for bootstrap percentile CIs, "basic"
for basic bootstrap CIs, and "norm" for normal approximation CIs.}

\item{...}{Further arguments to be passed to \code{boot::boot.ci()}.}
}
\description{
Calculate confidence intervals for chance-adjusted agreement indexes using
one of several approaches.
}
