% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensional_intraclass.R
\name{dim_icc}
\alias{dim_icc}
\title{Calculate intraclass correlation coefficients for reliability}
\usage{
dim_icc(
  .data,
  model = c("1A", "1B", "2", "2A", "3", "3A"),
  type = c("agreement", "consistency"),
  unit = c("single", "average"),
  object = Object,
  rater = Rater,
  score = Score,
  trial = NULL,
  bootstrap = 2000,
  warnings = TRUE
)
}
\arguments{
\item{.data}{Required. A data frame containing variables identifying and
quantifying each observation, in a "tall" format.}

\item{model}{Optional. A string indicating which ICC model to use, according
to Gwet's (2014) typology. Options include "1A", "1B", "2", "2A", "3", and
"3A". (default = \code{"1A"})}

\item{type}{Optional. A string indicating which ICC type to use, according to
McGraw & Wong's (1999) typology. Options include "agreement" and
"consistency". (default = \code{"agreement"})}

\item{unit}{Optional. A string indicating which unit of analysis to use.
Options include "single" for the reliability of a single randomly chosen
rater or "average" for the reliability of the average of all raters.
(default = \code{"single"})}

\item{object}{Optional. The name of the variable in \code{.data} identifying
the object of measurement for each observation, in non-standard evaluation
without quotation marks. (default = \code{Object})}

\item{rater}{Optional. The name of the variable in \code{.data} identifying
the rater or source of measurement for each observation, in non-standard
evaluation without quotation marks. (default = \code{Rater})}

\item{score}{Optional. The name of the variable in \code{.data} containing
the dimensional score or rating for each observation, in non-standard
evaluation without quotation marks. (default = \code{Score})}

\item{trial}{Optional. The name of the variable in \code{.data} identifying
the time or trial of measurement for each observation, in non-standard
evaluation without quotation marks. (default = \code{NULL})}

\item{bootstrap}{Optional. A non-negative integer indicating how many
bootstrap resamples to use in estimating confidence intervals. Set to
\code{0} to forgo bootstrapping. (default = \code{2000})}

\item{warnings}{Optional. A logical value indicating whether warnings should
be generated and displayed during this analysis. (default = \code{TRUE})}
}
\value{
A list object of class "agreement_icc" with the following elements:
\item{Object}{The estimated object variance} \item{Rater}{The estimated
rater variance} \item{Interaction}{The estimated object-by-variance
interaction variance} \item{Residual}{The estimated residual variance}
\item{Intra_ICC}{The estimated intra-rater reliability}
\item{Inter_ICC}{The estimated inter-rater reliability}
\item{boot_results}{The bootstrap results from the boot package}
\item{formulation}{The model, type, and k arguments} \item{details}{A list
containing counts and data from the analysis} \item{call}{The function call
that produced this output}
}
\description{
Analyze intra- and inter-rater reliability of dimensional data using various
intraclass correlation coefficients. These coefficients are implemented using
a generalized approach that accommodates missing data and multiple trials.
}
