% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensional_simulation.R
\name{dim_icc_sim}
\alias{dim_icc_sim}
\title{Simulate ICC by Number of Raters}
\usage{
dim_icc_sim(object, min_r = 1, max_r, plot = TRUE)
}
\arguments{
\item{object}{An object resulting from a call to \code{dim_icc()}.}

\item{min_r}{A nonnegative integer representing the minimum number of raters
to be simulated. (default = 1)}

\item{max_r}{A nonnegative integer representing the maximum number of raters
to be simulated.}

\item{plot}{A logical flag indicating whether to generate a simulation plot.
(default = TRUE)}
}
\value{
A tibble containing the estimated inter-rater ICC by number of
raters averaged.
}
\description{
Simulate ICC by Number of Raters
}
