% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{fish}
\alias{fish}
\title{Fish example data}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4 rows and 5 columns.}
\usage{
fish
}
\description{
Example categorical data describing 4 fish coded by 5 raters.
}
\keyword{datasets}
