% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{gwet3.3}
\alias{gwet3.3}
\title{Interval example data}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 16 rows and 4 columns.}
\usage{
gwet3.3
}
\description{
Example data with 16 objects rated by 4 raters.
}
\references{
Gwet (2014), Exercise 3.3
}
\keyword{datasets}
