% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_utilities.R
\name{long_to_wide}
\alias{long_to_wide}
\title{Convert long format to wide format}
\usage{
long_to_wide(x, rater, score, ...)
}
\arguments{
\item{x}{\emph{Required.} A data frame containing rating data in long format.}

\item{rater}{\emph{Required.} The name of the variable in \code{x} that indicates
the rater for each measurement.}

\item{score}{\emph{Required.} The name of the variable in \code{x} that indicates
the score for each measurement.}

\item{...}{\emph{Optional.} Additional arguments passed on to
\code{tidyr::pivot_wider()}.}
}
\value{
A tibble containing the data from \code{x} but in wide format.
}
\description{
Description
}
