% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{mcgraw}
\alias{mcgraw}
\title{McGraw example data}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 10 rows and 2 columns.}
\usage{
mcgraw
}
\description{
Example dimensional data describing 10 families and 2 family members.
}
\references{
\url{https://doi.org/10/br5ffs}
}
\keyword{datasets}
