% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{ordered_wide}
\alias{ordered_wide}
\title{Example dataset with ordered categorical ratings in wide format}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 20 rows and 5 columns.}
\usage{
ordered_wide
}
\description{
A tibble in wide format with variables denoting each rater. Example data with
20 objects rated by 5 raters into 4 categories denoted by the numbers
\code{0}, \code{1}, \code{2}, and \code{3}. Includes some missing data as
(\code{NA}). Adapted from Gwet (2014) Table 4.9 on page 125 as part of
Example 4.4.
}
\references{
Gwet, K. L. (2014). Handbook of inter-rater reliability: The
definitive guide to measuring the extent of agreement among raters (4th
ed.). Gaithersburg, MD: Advanced Analytics.
}
\seealso{
\code{\link{ordered}}
}
\keyword{datasets}
