\name{summary.agreement}
\alias{summary.agreement}
\title{
Summarize an Agreement Object
}
\description{
This function is a method for \code{summary} for an agreement object class \code{agreement}. 
}
\usage{
\method{summary}{agreement}(object, dec, \dots)
}
\arguments{
  \item{object}{
The name of an object of class \code{agreement}.
}
  \item{dec}{
Significant digits.
}
  \item{\dots}{
Arguments passed to the \code{print} method.
}
}
\author{
Yue Yu <yyu@imyy.net>
}
\seealso{
\code{\link{summary.report}}
}
