% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_utilities.R
\name{table_to_wide}
\alias{table_to_wide}
\title{Convert table format to wide format}
\usage{
table_to_wide(x, raters = NULL)
}
\arguments{
\item{x}{\emph{Required.} A matrix, dataframe, or table object representing a
two-by-two contingency matrix (AKA confusion matrix).}

\item{raters}{\emph{Optional.} A character vector containing the name of each
rater.}
}
\value{
A tibble containing the data from \code{x} in wide format.
}
\description{
Description
}
