% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_utilities.R
\name{wide_to_long}
\alias{wide_to_long}
\title{Convert wide format to long format}
\usage{
wide_to_long(x, ...)
}
\arguments{
\item{x}{\emph{Required.} A data frame containing rating data in wide format.}

\item{...}{\emph{Optional.} Additional arguments passed on to
\code{tidyr::pivot_longer()/}.}
}
\value{
A tibble containing the data from \code{x} but in long format.
}
\description{
Description
}
