\name{Agreement-package}
\alias{Agreement-package}
\alias{Agreement}
\docType{package}
\title{
Statistical Tools for Measuring Agreement
}
\description{
This package computes several statistics for measuring agreement, for example, mean square deviation (MSD), total deviation index (TDI) or concordance correlation coefficient (CCC). It can be used for both continuous data and categorical data for multiple raters and multiple readings cases.
}
\details{
\tabular{ll}{
Package: \tab Agreement\cr
Type: \tab Package\cr
Version: \tab 0.8-1\cr
Date: \tab 2011-03-05\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
}
}
\author{
Yue Yu AND Lawrence Lin

Maintainer: Yue Yu <yyu@imyy.net>
}
