% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agreement-package.R
\docType{package}
\name{agreement-package}
\alias{agreement}
\alias{agreement-package}
\title{agreement: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Jeffrey Girard \email{me@jmgirard.com}

}
\keyword{internal}
