% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefficients.R
\name{calc_all}
\alias{calc_all}
\title{Calculate all chance-adjusted indexes of categorical agreement}
\usage{
calc_all(.data, categories = NULL, weighting = c("identity", "linear",
  "quadratic"))
}
\arguments{
\item{.data}{Required.}

\item{categories}{Optional.}

\item{weighting}{Optional.}
}
\value{
A 5x4 tibble where each row corresponds to an index of categorical
agreement (i.e., S, gamma, kappa, pi, and) and the columns are the
following: \item{Approach}{The name of the coefficient}
\item{Observed}{Percent of agreement observed in the data (given the
coefficients' assumptions)} \item{Expected}{Percent of agreement expected
by chance (given the coefficients' assumptions)} \item{Adjusted}{The
estimated chance-adjusted agreement (i.e., the actual coefficient values)}
}
\description{
This function calculates
}
