% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specific.R
\name{calc_specific}
\alias{calc_specific}
\title{Calculate specific agreement from object-by-rater matrix}
\usage{
calc_specific(.data, categories = NULL)
}
\arguments{
\item{.data}{A numeric object-by-rater matrix where values are categories.}

\item{categories}{A numeric vector of possible categories. Useful when
\code{.data} may not contain all possible categories.}
}
\value{
A data frame containing the specific agreement coefficient for each
category.
}
\description{
Calculate the percent observed agreement for each category in an
object-by-rater matrix.
}
\examples{
data("fish")
calc_specific(fish, categories = 1:3)
}
