% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{unordered}
\alias{unordered}
\title{Example dataset with unordered categorical ratings from Gwet}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 12 rows and 4 columns.}
\usage{
unordered
}
\description{
An object-by-rater tibble containing category assignments. Example data with
12 objects rated by 4 raters into 5 unordered categories denoted by the
letters \code{a}, \code{b}, \code{c}, \code{d}, and \code{e}. Includes some
missing data \code{NA}. Adapted from Gwet (2014) Table 2.15 on page 55 as
part of Example 2.4.
}
\references{
Gwet, K. L. (2014). Handbook of inter-rater reliability: The
definitive guide to measuring the extent of agreement among raters (4th
ed.). Gaithersburg, MD: Advanced Analytics.
}
\keyword{datasets}
